Change long to int in calculation

+++ Main.java
@@ -4,12 +4,12 @@ public class Main {
         // Объявляете переменные для входных данных и
         // параметров программы: одну для хранения
         double ticketPrice = 1000.0;
-        System.out.println("Цена вашего билета " + (long) ticketPrice + " рублей " + (long) (ticketPrice - (long) ticketPrice) * 100 + " копеек ");
+        System.out.println("Цена вашего билета " + (int) ticketPrice + " рублей " + (int) (ticketPrice - (int) ticketPrice) * 100 + " копеек ");
 
         // стоимости билета, другую для хранения количества
         // рублей для одной бонусной милли
-        long bonusKoeficients = 20;
-        long bonusMiles = (long) (ticketPrice / bonusKoeficients);
+        int bonusKoeficients = 20;
+        int bonusMiles = (int) (ticketPrice / bonusKoeficients);
         System.out.println("Ваши бонусные мили " + bonusMiles);
 
         // Рассчитываете количество бонусных милль, используя
